+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,8 +21,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ReservationActivity" />
-        <activity android:name=".ReservationCompleteActivity" />
+        <activity android:name=".activity.ReservationActivity" />
+        <activity android:name=".activity.ReservationCompleteActivity" />
     </application>
 
 </manifest>
