+++ app/src/main/java/woowacourse/movie/adapter/RunningTimeSpinnerAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.adapter
 
 import android.view.View
 import android.view.ViewGroup
